Terminando index de la programacion de vehiculo
