+++ config/menu.php
@@ -57,7 +57,7 @@ Menu::macro('sidebar', function () {//COMIENZO DEL SIDEBAR EN VERSION DE MENU
 				          ->prepend('<a href="#"><i class="fas fa-truck-moving"></i> <span>'.trans('adminlte_lang::message.MenuVehicleTitle').'</span><i class="fas fa-angle-left pull-right" style="color:#FFFFFF;" width="18" height="18"></i></a>')
 				          ->addParentClass('treeview')
 				          ->add(Link::toUrl('/vehicle', '<i class="fas fa-list-alt"></i> '.trans('adminlte_lang::message.MenuVehiclelist')))
-				          ->add(Link::toUrl('/prueba', '<i class="fas fa-calendar-alt"></i> '.trans('adminlte_lang::message.MenuPrograVehic')))
+				          ->add(Link::toUrl('/vehicle-programacion', '<i class="fas fa-calendar-alt"></i> '.trans('adminlte_lang::message.MenuPrograVehic')))
 				          ->add(Link::toUrl('/vehicle-mantenimiento', '<i class="fas fa-tools"></i> '.trans('adminlte_lang::message.MenuMantVehic')))
 				          ->addClass('treeview-menu')
 			          )
